# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "Open Api",
        "contact": {
            "email": "admin@juzaweb.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/admin/post-type/{type}": {
            "get": {
                "tags": [
                    "Admin / Post Type"
                ],
                "summary": "Get list post type items",
                "operationId": "admin.post-type.type.index",
                "parameters": [
                    {
                        "name": "type",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "$ref": "#/components/parameters/query_limit"
                    },
                    {
                        "$ref": "#/components/parameters/query_keyword"
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/success_list"
                    },
                    "500": {
                        "$ref": "#/components/responses/error_500"
                    }
                }
            },
            "post": {
                "tags": [
                    "Post Type"
                ],
                "summary": "Create Post Type",
                "operationId": "admin.post-type.type.store",
                "parameters": [
                    {
                        "name": "type",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "Title",
                                        "type": "string"
                                    },
                                    "content": {
                                        "description": "content",
                                        "type": "string"
                                    },
                                    "grecaptcha_token": {
                                        "description": "(Optional) Token of Google Recaptcha V3",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "$ref": "#/components/responses/success_detail"
                    },
                    "422": {
                        "$ref": "#/components/responses/error_422"
                    },
                    "500": {
                        "$ref": "#/components/responses/error_500"
                    }
                }
            }
        },
        "/api/admin/post-type/{type}/{id}": {
            "get": {
                "tags": [
                    "Post Type"
                ],
                "summary": "Get post type data by id",
                "operationId": "admin.post-type.type.show",
                "parameters": [
                    {
                        "name": "type",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "$ref": "#/components/parameters/id_in_path"
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/success_detail"
                    },
                    "404": {
                        "$ref": "#/components/responses/error_404"
                    },
                    "500": {
                        "$ref": "#/components/responses/error_500"
                    }
                }
            },
            "put": {
                "tags": [
                    "Post Type"
                ],
                "summary": "Update Post Type",
                "operationId": "admin.post-type.type.update",
                "parameters": [
                    {
                        "name": "type",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "Title",
                                        "type": "string"
                                    },
                                    "content": {
                                        "description": "content",
                                        "type": "string"
                                    },
                                    "grecaptcha_token": {
                                        "description": "(Optional) Token of Google Recaptcha V3",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/success_detail"
                    },
                    "422": {
                        "$ref": "#/components/responses/error_422"
                    },
                    "500": {
                        "$ref": "#/components/responses/error_500"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Post Type"
                ],
                "summary": "Delete Post Type",
                "operationId": "admin.post-type.type.destroy",
                "parameters": [
                    {
                        "name": "type",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/success_detail"
                    },
                    "422": {
                        "$ref": "#/components/responses/error_422"
                    },
                    "500": {
                        "$ref": "#/components/responses/error_500"
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "User login",
                "operationId": "api.auth.login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "email",
                                        "type": "string",
                                        "example": "string@gmail.com"
                                    },
                                    "password": {
                                        "description": "password",
                                        "type": "string",
                                        "example": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "$ref": "#/components/responses/success_detail"
                    },
                    "422": {
                        "$ref": "#/components/responses/error_422"
                    },
                    "500": {
                        "$ref": "#/components/responses/error_500"
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "User logout",
                "operationId": "api.auth.logout",
                "responses": {
                    "201": {
                        "$ref": "#/components/responses/success_detail"
                    },
                    "422": {
                        "$ref": "#/components/responses/error_422"
                    },
                    "500": {
                        "$ref": "#/components/responses/error_500"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "responses": {
            "success_detail": {
                "description": "Get Data Success",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "data": {
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "success_list": {
                "description": "Get List Success",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
            "error_404": {
                "description": "Page not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "errors": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "code": {
                                                "type": "string",
                                                "example": ""
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": ""
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "message": {
                                    "type": "string",
                                    "example": ""
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "error_422": {
                "description": "Validate Error",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "errors": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "field": {
                                                "type": "string",
                                                "example": ""
                                            },
                                            "message": {
                                                "type": "string",
                                                "example": ""
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "message": {
                                    "type": "string",
                                    "example": ""
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "error_500": {
                "description": "Server Error",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "errors": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "code": {
                                                "type": "string",
                                                "example": ""
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": ""
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "message": {
                                    "type": "string",
                                    "example": ""
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "parameters": {
            "id_in_path": {
                "name": "id",
                "in": "path",
                "required": true,
                "schema": {
                    "type": "string"
                }
            },
...
            "query_limit": {
                "name": "limit",
                "in": "query",
                "schema": {
                    "type": "integer"
                }
            },
...
            "query_keyword": {
                "name": "keyword",
                "in": "query",
                "schema": {
                    "type": "string"
                }
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer :token)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}